Presentation of expression QTL densities
The distribution of expression QTL around chromosomes differs
with the minimal LOD score (or arbitrary other filters) that
is requested from the eQTLs. This script provides a routine
to plot an image that represents both at the same time:
vertically the chromosomal location, with its intensities
the number of QTLs that are found at a particular region are represented
and horizontally the variation with a change of the threshold
are interpretable.