Improved abstractions for other projects
To help the matching with numerical IDs,
the matching against ^MT of the identifiers
was substituted with an abstract matching
against the template ^INDIVIDUALIDPREFIX.

Directory separators were introduced for the
specification of the path to the genotype
data.

The PHENOCOL attribute now also influences
the preparation of the R/qtl input file.